A tiny bit of work

+++ src/main/java/com/goosejs/apollo/client/renderer/font/TrueTypeFontRenderer.java
@@ -83,7 +83,7 @@ public class TrueTypeFontRenderer
         this.additionFontSpacing = additionalFontSpacing;
         fontShader = new FontShader();
         loadFonts(fontFile);
-        GlobalPerspectiveMatrices.add2DPerspectiveOnChange((x02, y02, x12, y12) -> windowResized(x12, y12));
+        GlobalPerspectiveMatrices.add2DPerspectiveOnChange((x02, y02, x12, y12) -> windowResized((int)x12, (int)y12));
         this.originalX = x0;
         this.originalWidth = x1;
         this.originalY = y0;
